Improve installation of non-packaged programs

+++ install-non-packaged-programs.sh
@@ -4,29 +4,49 @@
 # package repo, but instead installed via some language's package manager, such
 # as npm or cargo.
 
-#if [[ $# -lt 1 ]]; then
-#    echo "Usage: install-pkg-fedora.sh <package-list>..."
-#    echo "  <package-list> is a file containing one package name per line"
-#    exit 1
-#fi
+echo_status() {
+    echo -e "\033[1m\033[32m***\033[94m" $* "\033[0m"
+}
 
-if [ $(id -u) = 0 ]; then
-    echo "This script shouldn't be run as root."
+echo_error() {
+    echo -e "\033[1m\033[31m!!!\033[31m" $* "\033[0m"
+}
+
+if [[ $# -gt 0 ]]; then
+    echo "This script will check for/install the following:"
+    echo ""
+#    echo "   - Joplin"
+    echo "   - brew (Homebrew)"
+    echo "   - grex"
+    echo "   - jupyter-lab"
+    echo "   - pet"
+    echo "   - starship"
+    echo "   - timg"
+    echo "   - virtualfish"
+    echo "   - xcape"
+    echo "   - yq"
+    echo ""
+    echo_error "Run this script without arguments to execute these actions."
     exit 1
 fi
 
-if ! [ -f ~/.joplin/Joplin.AppImage ]; then
-    echo ">>> Installing Joplin for Desktop [via custom installer]"
-    wget -O - https://raw.githubusercontent.com/laurent22/joplin/dev/Joplin_install_and_update.sh | bash
+if [ $(id -u) = 0 ]; then
+    echo_error "This script shouldn't be run as root."
+    exit 1
 fi
 
+#if ! [ -f ~/.joplin/Joplin.AppImage ]; then
+#    echo_status "Installing Joplin for Desktop [via custom installer]"
+#    wget -O - https://raw.githubusercontent.com/laurent22/joplin/dev/Joplin_install_and_update.sh | bash
+#fi
+
 if ! which starship >/dev/null; then
-    echo ">>> Installing starship [via starship.rs/install.sh]"
+    echo_status "Installing starship [via starship.rs/install.sh]"
     curl -sS https://starship.rs/install.sh | sh
 fi
 
 if ! which xcape >/dev/null; then
-    echo ">>> Installing xcape [via github.com/alols/xcape]"
+    echo_status "Installing xcape [via github.com/alols/xcape]"
     sudo dnf install git gcc make pkgconfig libX11-devel libXtst-devel libXi-devel
     THISDIR=$(pwd)
     cd /tmp
@@ -38,23 +58,22 @@ if ! which xcape >/dev/null; then
     sudo make install
     cd $THISDIR
     if which xcape >/dev/null; then
-        echo ">>> Successfully installed xcape"
         rm -rf /tmp/xcape
     else
-        echo "!!! Installation of xcape failed"
+        echo_error "Installation of xcape failed"
     fi
 fi
 
 if ! which pet >/dev/null; then
     if ! which jq >/dev/null; then
-        echo "!!! Installation of pet requires jq; skipping"
+        echo_error "Installation of pet requires jq; skipping"
     else
-        echo ">>> Installing pet [via github.com/knqyf263/pet]"
+        echo_status "Installing pet [via github.com/knqyf263/pet]"
         petrpm=$(curl -s https://api.github.com/repos/knqyf263/pet/releases/latest | jq '.assets[] | select(.name|match("linux_amd64.rpm$")) | .browser_download_url' | tr -d \")
         if [[ -z "$petrpm" ]]; then
-            echo "!!! Installation of pet failed: no linux_amd64.rpm found (maybe the latest release doesn't have one?)"
+            echo_error "Installation of pet failed: no linux_amd64.rpm found (maybe the latest release doesn't have one?)"
         else
-            sudo dnf install -y "$petrpm" && echo ">>> Successfully installed pet"
+            sudo dnf install -y "$petrpm"
         fi
     fi
 fi
@@ -67,14 +86,41 @@ if which npm >/dev/null; then
     fi
     export NPM_CONFIG_PREFIX=~/.npm-global
 else
-    echo "!!! npm isn't installed; skipping installation of npm packages."
+    echo_error "NPM isn't installed; skipping installation of npm packages."
 fi
 
 if which python >/dev/null; then
     python -m pip install --upgrade pip
-    python -m pip install --user virtualfish
-    python -m pip install --user docopt black rich textual
-    python -m pip install --user jupyterlab
+    if ! [[ -e ~/.local/bin/vf ]]; then
+        echo_status "Installing virtualfish [via pip]"
+        python -m pip install --user virtualfish
+    fi
+    if ! [[ -e ~/.local/bin/jupyter-lab ]]; then
+        echo_status "Installing jupyter-lab [via pip]"
+        python -m pip install --user jupyterlab
+    fi
 else
-    echo "!!! python isn't installed; skipping installation of Python packages."
+    echo_error "Python isn't installed; skipping installation of Python packages."
+fi
+
+if ! which brew >/dev/null; then
+    echo_status "Installing Homebrew"
+    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+fi
+
+if ! which brew >/dev/null; then
+    echo_error "Homebrew isn't installed; skipping installation of Homebrew packages."
+else
+    if ! [[ -e /home/linuxbrew/.linuxbrew/bin/yq ]]; then
+        echo_status "Installing yq [via brew]"
+        brew install yq
+    fi
+    if ! [[ -e /home/linuxbrew/.linuxbrew/bin/timg ]]; then
+        echo_status "Installing timg [via brew]"
+        brew install timg
+    fi
+    if ! [[ -e /home/linuxbrew/.linuxbrew/bin/grex ]]; then
+        echo_status "Installing grex [via brew]"
+        brew install grex
+    fi
 fi
